refactor[w4.1]:move logic about jwt in controller
