+++ HealthcareSystem/PatientWebApp/Controllers/UserController.cs
@@ -1,9 +1,16 @@
-﻿using Backend.Model.Exceptions;
+﻿using System;
+using System.Text;
+using Backend.Model.Exceptions;
 using Backend.Service;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Model.Users;
 using PatientWebApp.DTOs;
+using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using Microsoft.Extensions.Configuration;
+using System.Collections.Generic;
 
 namespace PatientWebApp.Controllers
 {
@@ -12,12 +19,12 @@ namespace PatientWebApp.Controllers
     [ApiController]
     public class UserController : ControllerBase
     {
-        private readonly IJWTAuthenticationManager _jWTAuthenticationManager;
+        private readonly IConfiguration _config;
         private readonly IPatientService _patientService;
         private readonly IAdminService _adminService;
-        public UserController(IJWTAuthenticationManager jWTAuthenticationManager, IPatientService patientService, IAdminService adminService)
+        public UserController(IConfiguration config, IPatientService patientService, IAdminService adminService)
         {
-            _jWTAuthenticationManager = jWTAuthenticationManager;
+            _config = config;
             _patientService = patientService;
             _adminService = adminService;
         }
@@ -49,13 +56,36 @@ namespace PatientWebApp.Controllers
                 return StatusCode(500, exception.Message);
             }
         }
+
+        //This method will be corrected by a colleague Jelena Budisa
+        /*
+          public LoggedUserDTO GetLoggedUser(string token)
+        {
+            var key = Encoding.ASCII.GetBytes(_tokenKey);
+            var handler = new JwtSecurityTokenHandler();
+            var validations = new TokenValidationParameters
+            {
+                ValidateIssuerSigningKey = true,
+                IssuerSigningKey = new SymmetricSecurityKey(key),
+                ValidateIssuer = false,
+                ValidateAudience = false
+            };
+            var claims = handler.ValidateToken(token, validations, out var tokenSecure);
+
+            string username = claims.Claims.First(c => c.Type == ClaimTypes.NameIdentifier).Value;
+            string role = claims.Claims.First(c => c.Type == ClaimTypes.Role).Value;
+            string jmbg = claims.Claims.First(c => c.Type == "Jmbg").Value;
+
+            return new LoggedUserDTO(username, jmbg, role);
+        }*/
+
         private string TryToLoginPatient(string username, string password)
         {
             Patient patient;
             try
             {
                 patient = _patientService.GetPatientByUsernameAndPassword(username, password);
-                string token = _jWTAuthenticationManager.Authenticate(patient.Username, patient.Jmbg, UserRoles.Patient);
+                string token = GenerateJWT(patient.Username, patient.Jmbg, UserRoles.Patient);
                 return token;
             }
             catch (NotFoundException)
@@ -69,13 +99,33 @@ namespace PatientWebApp.Controllers
             try
             {
                 admin = _adminService.GetAdminByUsernameAndPassword(username, password);
-                string token = _jWTAuthenticationManager.Authenticate(admin.Username, admin.Jmbg, UserRoles.Admin);
+                string token = GenerateJWT(admin.Username, admin.Jmbg, UserRoles.Admin);
                 return token;
             }
             catch (NotFoundException)
             {
                 return null;
             }
+        }
+
+        private string GenerateJWT(string username, string jmbg, string role)
+        {
+            var tokenKey = _config.GetValue<string>("TokenKey");
+            var key = Encoding.ASCII.GetBytes(tokenKey);
+            var tokenHandler = new JwtSecurityTokenHandler();
+            var tokenDescriptor = new SecurityTokenDescriptor
+            {
+                Subject = new ClaimsIdentity(new Claim[]
+                {
+                    new Claim(ClaimTypes.NameIdentifier, username), new Claim("Jmbg",jmbg) , new Claim(ClaimTypes.Role, role)
+                }),
+                Expires = DateTime.UtcNow.AddHours(1),
+                SigningCredentials = new SigningCredentials(
+                    new SymmetricSecurityKey(key),
+                    SecurityAlgorithms.HmacSha256Signature)
+            };
+            var token = tokenHandler.CreateToken(tokenDescriptor);
+            return tokenHandler.WriteToken(token);
         }
     }
 }
