+++ HealthcareSystem/PatientWebApp/Startup.cs
@@ -193,7 +193,6 @@ namespace PatientWebApp
                 };
             });
 
-            services.AddSingleton<IJWTAuthenticationManager>(new JWTAuthenticationManager(tokenKey));
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
