+++ HealthcareSystem/PatientWebApp/appsettings.json
@@ -23,5 +23,5 @@
     "RetryCount": 5,
     "RetryWaitInSeconds": 30
   },
-  "TokenKey": "This is my test private key"
+  "TokenKey": "This is my private key"
 }
